install Bootstrap UI, user input fix, unit choice for text

+++ db/migrate/20130402120548_create_texts.rb
@@ -8,7 +8,7 @@ class CreateTexts < ActiveRecord::Migration
       t.integer :lexiles
       t.float :sequence
       t.float :genre
-      t.integer :performance
+      t.float :performance
       t.timestamps
     end
   end
